refactor(Menu, Product): infra에 위치한 ProfanityClients 인터페이스를 domain 계층 내부로 이동
