+++ src/main/java/kitchenpos/menus/tobe/domain/vo/ProfanityClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.infra;
+package kitchenpos.menus.tobe.domain.vo;
 
 public interface ProfanityClient {
     boolean containsProfanity(String text);
