+++ src/main/java/kitchenpos/products/tobe/domain/vo/ProfanityClient.java
@@ -1,0 +1,5 @@
+package kitchenpos.products.tobe.domain.vo;
+
+public interface ProfanityClient {
+    boolean containsProfanity(String text);
+}
